refactor: split tcop postgres module into submodules
